extract AVIF and AVC boxes into their own files
